added handling fo first deployment no compose file exists yet

+++ FylumDeploy.ComposeBuilder/PodmanCompose/PodmanComposeService.cs
@@ -2,17 +2,28 @@
 
 namespace FylumDeploy.ComposeBuilder.PodmanCompose;
 
-internal class PodmanComposeService : IPodmanComposeService
+internal class PodmanComposeService(
+    ILogger<PodmanComposeService> logger,
+    IProcessExecutionService processExecutionService)
+    : IPodmanComposeService
 {
-    private readonly IProcessExecutionService _processExecutionService;
-
-    public PodmanComposeService(IProcessExecutionService processExecutionService)
-    {
-        _processExecutionService = processExecutionService;
-    }
+    private readonly ILogger<PodmanComposeService> _logger = logger;
+    private readonly IProcessExecutionService _processExecutionService = processExecutionService;
 
     public async Task<bool> ComposeDownAsync(CancellationToken cancellationToken)
     {
+        string workingDirectory = Directories.OutputDirectory;
+        string composeFilePath = Path.Combine(workingDirectory, "podman-compose.yaml");
+
+        // 1. Prüfen, ob dies ein Update oder das erste Deployment ist
+        if (!File.Exists(composeFilePath))
+        {
+            _logger.LogWarning("No previous deployment found. Skipping 'compose down'");
+            return true;
+        }
+
+        _logger.LogInformation("Existing deployment found. Trying to stop old containers...");
+
         var command = $"podman compose down";
         var processExecute = new ProcessExecute(command: command, workingDirectory: Directories.OutputDirectory);
         var result = await _processExecutionService.ExecuteProcessAsync(processExecute, cancellationToken);
@@ -20,6 +31,7 @@ internal class PodmanComposeService : IPodmanComposeService
     }
     public async Task<bool> ComposeUpAsync(CancellationToken cancellationToken)
     {
+        _logger.LogInformation("Composing up new containers...");
         var command = $"podman compose up -d";
         var processExecute = new ProcessExecute(command: command, workingDirectory: Directories.OutputDirectory);
         var result = await _processExecutionService.ExecuteProcessAsync(processExecute, cancellationToken);
